User can now create a new Expense
- Added NewExpenseScreen which is a form for the new Expense
- the FAB opens the NewExpenseScreen
- made Budget a ChangeNotifier
- made BudgetInfoWidget a child of ChangeNotifierProvider with value
  constructor of its corresponding budget.
- realised that TripTab doesn't need to be statful because its state
  depends on the ChangeNotifierProvider only

+++ lib/models/expense.dart
@@ -2,7 +2,7 @@ import 'package:wisebud/models/budget.dart';
 import 'package:wisebud/models/trip.dart';
 
 class Expense {
-  int amount;
+  double amount;
   late String desc;
   late DateTime time;
   late String currency; // make it an enum instead of String
